test: move LSP6 + LSP20 reentrancy tests in separate test suite
