+++ tests/LSP20CallVerification/LSP6/Interactions/PermissionTransferValue.test.ts
@@ -1,7 +1,6 @@
 import { expect } from "chai";
 import { ethers } from "hardhat";
 import { SignerWithAddress } from "@nomiclabs/hardhat-ethers/signers";
-import { EIP191Signer } from "@lukso/eip191-signer.js";
 import { BigNumber } from "ethers";
 import {
   ExecutorLSP20,
