+++ tests/Reentrancy/LSP20/ERC725XExecuteToLSP6BatchExecuteRelayCall.test.ts
@@ -3,12 +3,13 @@ import { ethers } from "hardhat";
 
 //types
 import { BigNumber, BytesLike } from "ethers";
+import { SingleReentrancyRelayer__factory } from "../../../types";
 
 // constants
-import { ALL_PERMISSIONS, ERC725YDataKeys } from "../../../../constants";
+import { ERC725YDataKeys, OPERATION_TYPES } from "../../../constants";
 
 // setup
-import { LSP6TestContext } from "../../../utils/context";
+import { LSP6TestContext } from "../../utils/context";
 
 // helpers
 import {
@@ -22,11 +23,11 @@ import {
   addUniversalReceiverDelegateTestCases,
   changeUniversalReceiverDelegateTestCases,
   // Functions
+  generateBatchRelayPayload,
   loadTestCase,
 } from "./reentrancyHelpers";
-import { LSP20ReentrantContract__factory } from "../../../../types";
 
-export const testERC725XExecuteToERC725XExecute = (
+export const testERC725XExecuteToLSP6BatchExecuteRelayCall = (
   buildContext: (initialFunding?: BigNumber) => Promise<LSP6TestContext>,
   buildReentrancyContext: (
     context: LSP6TestContext
@@ -34,37 +35,46 @@ export const testERC725XExecuteToERC725XExecute = (
 ) => {
   let context: LSP6TestContext;
   let reentrancyContext: ReentrancyContext;
+  let executeCalldata: {
+    operationType: number;
+    to: string;
+    value: number;
+    data: BytesLike;
+  };
 
   before(async () => {
     context = await buildContext(ethers.utils.parseEther("10"));
     reentrancyContext = await buildReentrancyContext(context);
+
+    const reentrantCall =
+      new SingleReentrancyRelayer__factory().interface.encodeFunctionData(
+        "relayCallThatReenters",
+        [context.keyManager.address]
+      );
+
+    executeCalldata = {
+      operationType: OPERATION_TYPES.CALL,
+      to: reentrancyContext.batchReentarncyRelayer.address,
+      value: 0,
+      data: reentrantCall,
+    };
   });
 
   describe("when reentering and transferring value", () => {
-    let executeCalldata: {
-      operationType: number;
-      to: string;
-      value: number;
-      data: BytesLike;
-    };
-
     before(async () => {
-      const reentrantCall =
-        new LSP20ReentrantContract__factory().interface.encodeFunctionData(
-          "callThatReenters",
-          ["TRANSFERVALUE"]
-        );
-
-      executeCalldata = {
-        operationType: 0,
-        to: reentrancyContext.reentrantContract.address,
-        value: 0,
-        data: reentrantCall,
-      };
+      await generateBatchRelayPayload(
+        context.universalProfile,
+        context.keyManager,
+        "TRANSFERVALUE",
+        reentrancyContext.batchReentarncyRelayer,
+        reentrancyContext.reentrantSigner,
+        reentrancyContext.newControllerAddress,
+        reentrancyContext.newURDAddress
+      );
     });
 
     transferValueTestCases.NotAuthorised.forEach((testCase) => {
-      it(`should revert if the reentrant contract has the following permission set: PRESENT - ${
+      it(`should revert if the reentrant signer has the following permission set: PRESENT - ${
         testCase.permissionsText
       }; MISSING - ${testCase.missingPermission}; AllowedCalls - ${
         testCase.allowedCalls ? "YES" : "NO"
@@ -73,8 +83,8 @@ export const testERC725XExecuteToERC725XExecute = (
           "TRANSFERVALUE",
           testCase,
           context,
-          reentrancyContext.reentrantContract.address,
-          reentrancyContext.reentrantContract.address
+          reentrancyContext.reentrantSigner.address,
+          reentrancyContext.batchReentarncyRelayer.address
         );
 
         await expect(
@@ -89,19 +99,19 @@ export const testERC725XExecuteToERC725XExecute = (
         )
           .to.be.revertedWithCustomError(context.keyManager, "NotAuthorised")
           .withArgs(
-            reentrancyContext.reentrantContract.address,
+            reentrancyContext.reentrantSigner.address,
             testCase.missingPermission
           );
       });
     });
 
-    it("should revert if the reentrant contract has the following permissions: REENTRANCY, TRANSFERVALUE & NO AllowedCalls", async () => {
+    it("should revert if the reentrant signer has the following permissions: REENTRANCY, TRANSFERVALUE & NO AllowedCalls", async () => {
       await loadTestCase(
         "TRANSFERVALUE",
         transferValueTestCases.NoCallsAllowed,
         context,
-        reentrancyContext.reentrantContract.address,
-        reentrancyContext.reentrantContract.address
+        reentrancyContext.reentrantSigner.address,
+        reentrancyContext.batchReentarncyRelayer.address
       );
 
       await expect(
@@ -116,13 +126,13 @@ export const testERC725XExecuteToERC725XExecute = (
       ).to.be.revertedWithCustomError(context.keyManager, "NoCallsAllowed");
     });
 
-    it("should pass if the reentrant contract has the following permissions: REENTRANCY, TRANSFERVALUE & AllowedCalls", async () => {
+    it("should pass if the reentrant signer has the following permissions: REENTRANCY, TRANSFERVALUE & AllowedCalls", async () => {
       await loadTestCase(
         "TRANSFERVALUE",
         transferValueTestCases.ValidCase,
         context,
-        reentrancyContext.reentrantContract.address,
-        reentrancyContext.reentrantContract.address
+        reentrancyContext.reentrantSigner.address,
+        reentrancyContext.batchReentarncyRelayer.address
       );
 
       expect(
@@ -148,37 +158,27 @@ export const testERC725XExecuteToERC725XExecute = (
 
       expect(
         await context.universalProfile.provider.getBalance(
-          reentrancyContext.reentrantContract.address
+          reentrancyContext.batchReentarncyRelayer.address
         )
       ).to.equal(ethers.utils.parseEther("1"));
     });
   });
 
   describe("when reentering and setting data", () => {
-    let executeCalldata: {
-      operationType: number;
-      to: string;
-      value: number;
-      data: BytesLike;
-    };
-
     before(async () => {
-      const reentrantCall =
-        new LSP20ReentrantContract__factory().interface.encodeFunctionData(
-          "callThatReenters",
-          ["SETDATA"]
-        );
-
-      executeCalldata = {
-        operationType: 0,
-        to: reentrancyContext.reentrantContract.address,
-        value: 0,
-        data: reentrantCall,
-      };
+      await generateBatchRelayPayload(
+        context.universalProfile,
+        context.keyManager,
+        "SETDATA",
+        reentrancyContext.batchReentarncyRelayer,
+        reentrancyContext.reentrantSigner,
+        reentrancyContext.newControllerAddress,
+        reentrancyContext.newURDAddress
+      );
     });
 
     setDataTestCases.NotAuthorised.forEach((testCase) => {
-      it(`should revert if the reentrant contract has the following permission set: PRESENT - ${
+      it(`should revert if the reentrant signer has the following permission set: PRESENT - ${
         testCase.permissionsText
       }; MISSING - ${testCase.missingPermission}; AllowedERC725YDataKeys - ${
         testCase.allowedERC725YDataKeys ? "YES" : "NO"
@@ -187,8 +187,8 @@ export const testERC725XExecuteToERC725XExecute = (
           "SETDATA",
           testCase,
           context,
-          reentrancyContext.reentrantContract.address,
-          reentrancyContext.reentrantContract.address
+          reentrancyContext.reentrantSigner.address,
+          reentrancyContext.batchReentarncyRelayer.address
         );
 
         await expect(
@@ -203,19 +203,19 @@ export const testERC725XExecuteToERC725XExecute = (
         )
           .to.be.revertedWithCustomError(context.keyManager, "NotAuthorised")
           .withArgs(
-            reentrancyContext.reentrantContract.address,
+            reentrancyContext.reentrantSigner.address,
             testCase.missingPermission
           );
       });
     });
 
-    it("should revert if the reentrant contract has the following permissions: REENTRANCY, SETDATA & NO AllowedERC725YDataKeys", async () => {
+    it("should revert if the reentrant signer has the following permissions: REENTRANCY, SETDATA & NO AllowedERC725YDataKeys", async () => {
       await loadTestCase(
         "SETDATA",
         setDataTestCases.NoERC725YDataKeysAllowed,
         context,
-        reentrancyContext.reentrantContract.address,
-        reentrancyContext.reentrantContract.address
+        reentrancyContext.reentrantSigner.address,
+        reentrancyContext.batchReentarncyRelayer.address
       );
 
       await expect(
@@ -233,13 +233,13 @@ export const testERC725XExecuteToERC725XExecute = (
       );
     });
 
-    it("should pass if the reentrant contract has the following permissions: REENTRANCY, SETDATA & AllowedERC725YDataKeys", async () => {
+    it("should pass if the reentrant signer has the following permissions: REENTRANCY, SETDATA & AllowedERC725YDataKeys", async () => {
       await loadTestCase(
         "SETDATA",
         setDataTestCases.ValidCase,
         context,
-        reentrancyContext.reentrantContract.address,
-        reentrancyContext.reentrantContract.address
+        reentrancyContext.reentrantSigner.address,
+        reentrancyContext.batchReentarncyRelayer.address
       );
 
       await context.universalProfile
@@ -265,36 +265,26 @@ export const testERC725XExecuteToERC725XExecute = (
   });
 
   describe("when reentering and adding permissions", () => {
-    let executeCalldata: {
-      operationType: number;
-      to: string;
-      value: number;
-      data: BytesLike;
-    };
-
     before(async () => {
-      const reentrantCall =
-        new LSP20ReentrantContract__factory().interface.encodeFunctionData(
-          "callThatReenters",
-          ["ADDCONTROLLER"]
-        );
-
-      executeCalldata = {
-        operationType: 0,
-        to: reentrancyContext.reentrantContract.address,
-        value: 0,
-        data: reentrantCall,
-      };
+      await generateBatchRelayPayload(
+        context.universalProfile,
+        context.keyManager,
+        "ADDCONTROLLER",
+        reentrancyContext.batchReentarncyRelayer,
+        reentrancyContext.reentrantSigner,
+        reentrancyContext.newControllerAddress,
+        reentrancyContext.newURDAddress
+      );
     });
 
     addPermissionsTestCases.NotAuthorised.forEach((testCase) => {
-      it(`should revert if the reentrant contract has the following permission set: PRESENT - ${testCase.permissionsText}; MISSING - ${testCase.missingPermission};`, async () => {
+      it(`should revert if the reentrant signer has the following permission set: PRESENT - ${testCase.permissionsText}; MISSING - ${testCase.missingPermission};`, async () => {
         await loadTestCase(
           "ADDCONTROLLER",
           testCase,
           context,
-          reentrancyContext.reentrantContract.address,
-          reentrancyContext.reentrantContract.address
+          reentrancyContext.reentrantSigner.address,
+          reentrancyContext.batchReentarncyRelayer.address
         );
 
         await expect(
@@ -309,19 +299,19 @@ export const testERC725XExecuteToERC725XExecute = (
         )
           .to.be.revertedWithCustomError(context.keyManager, "NotAuthorised")
           .withArgs(
-            reentrancyContext.reentrantContract.address,
+            reentrancyContext.reentrantSigner.address,
             testCase.missingPermission
           );
       });
     });
 
-    it("should pass if the reentrant contract has the following permissions: REENTRANCY, ADDCONTROLLER", async () => {
+    it("should pass if the reentrant signer has the following permissions: REENTRANCY, ADDCONTROLLER", async () => {
       await loadTestCase(
         "ADDCONTROLLER",
         addPermissionsTestCases.ValidCase,
         context,
-        reentrancyContext.reentrantContract.address,
-        reentrancyContext.reentrantContract.address
+        reentrancyContext.reentrantSigner.address,
+        reentrancyContext.batchReentarncyRelayer.address
       );
 
       await context.universalProfile
@@ -336,46 +326,38 @@ export const testERC725XExecuteToERC725XExecute = (
       const hardcodedPermissionKey =
         ERC725YDataKeys.LSP6["AddressPermissions:Permissions"] +
         reentrancyContext.newControllerAddress.substring(2);
+      const hardcodedPermissionValue =
+        "0x0000000000000000000000000000000000000000000000000000000000000010";
 
       expect(
         await context.universalProfile["getData(bytes32)"](
           hardcodedPermissionKey
         )
-      ).to.equal(ALL_PERMISSIONS);
+      ).to.equal(hardcodedPermissionValue);
     });
   });
 
   describe("when reentering and changing permissions", () => {
-    let executeCalldata: {
-      operationType: number;
-      to: string;
-      value: number;
-      data: BytesLike;
-    };
-
     before(async () => {
-      const reentrantCall =
-        new LSP20ReentrantContract__factory().interface.encodeFunctionData(
-          "callThatReenters",
-          ["EDITPERMISSIONS"]
-        );
-
-      executeCalldata = {
-        operationType: 0,
-        to: reentrancyContext.reentrantContract.address,
-        value: 0,
-        data: reentrantCall,
-      };
+      await generateBatchRelayPayload(
+        context.universalProfile,
+        context.keyManager,
+        "EDITPERMISSIONS",
+        reentrancyContext.batchReentarncyRelayer,
+        reentrancyContext.reentrantSigner,
+        reentrancyContext.newControllerAddress,
+        reentrancyContext.newURDAddress
+      );
     });
 
     editPermissionsTestCases.NotAuthorised.forEach((testCase) => {
-      it(`should revert if the reentrant contract has the following permission set: PRESENT - ${testCase.permissionsText}; MISSING - ${testCase.missingPermission};`, async () => {
+      it(`should revert if the reentrant signer has the following permission set: PRESENT - ${testCase.permissionsText}; MISSING - ${testCase.missingPermission};`, async () => {
         await loadTestCase(
           "EDITPERMISSIONS",
           testCase,
           context,
-          reentrancyContext.reentrantContract.address,
-          reentrancyContext.reentrantContract.address
+          reentrancyContext.reentrantSigner.address,
+          reentrancyContext.batchReentarncyRelayer.address
         );
 
         await expect(
@@ -390,19 +372,19 @@ export const testERC725XExecuteToERC725XExecute = (
         )
           .to.be.revertedWithCustomError(context.keyManager, "NotAuthorised")
           .withArgs(
-            reentrancyContext.reentrantContract.address,
+            reentrancyContext.reentrantSigner.address,
             testCase.missingPermission
           );
       });
     });
 
-    it("should pass if the reentrant contract has the following permissions: REENTRANCY, EDITPERMISSIONS", async () => {
+    it("should pass if the reentrant signer has the following permissions: REENTRANCY, EDITPERMISSIONS", async () => {
       await loadTestCase(
         "EDITPERMISSIONS",
         editPermissionsTestCases.ValidCase,
         context,
-        reentrancyContext.reentrantContract.address,
-        reentrancyContext.reentrantContract.address
+        reentrancyContext.reentrantSigner.address,
+        reentrancyContext.batchReentarncyRelayer.address
       );
 
       await context.universalProfile
@@ -428,36 +410,26 @@ export const testERC725XExecuteToERC725XExecute = (
   });
 
   describe("when reentering and adding URD", () => {
-    let executeCalldata: {
-      operationType: number;
-      to: string;
-      value: number;
-      data: BytesLike;
-    };
-
     before(async () => {
-      const reentrantCall =
-        new LSP20ReentrantContract__factory().interface.encodeFunctionData(
-          "callThatReenters",
-          ["ADDUNIVERSALRECEIVERDELEGATE"]
-        );
-
-      executeCalldata = {
-        operationType: 0,
-        to: reentrancyContext.reentrantContract.address,
-        value: 0,
-        data: reentrantCall,
-      };
+      await generateBatchRelayPayload(
+        context.universalProfile,
+        context.keyManager,
+        "ADDUNIVERSALRECEIVERDELEGATE",
+        reentrancyContext.batchReentarncyRelayer,
+        reentrancyContext.reentrantSigner,
+        reentrancyContext.newControllerAddress,
+        reentrancyContext.newURDAddress
+      );
     });
 
     addUniversalReceiverDelegateTestCases.NotAuthorised.forEach((testCase) => {
-      it(`should revert if the reentrant contract has the following permission set: PRESENT - ${testCase.permissionsText}; MISSING - ${testCase.missingPermission};`, async () => {
+      it(`should revert if the reentrant signer has the following permission set: PRESENT - ${testCase.permissionsText}; MISSING - ${testCase.missingPermission};`, async () => {
         await loadTestCase(
           "ADDUNIVERSALRECEIVERDELEGATE",
           testCase,
           context,
-          reentrancyContext.reentrantContract.address,
-          reentrancyContext.reentrantContract.address
+          reentrancyContext.reentrantSigner.address,
+          reentrancyContext.batchReentarncyRelayer.address
         );
 
         await expect(
@@ -472,19 +444,19 @@ export const testERC725XExecuteToERC725XExecute = (
         )
           .to.be.revertedWithCustomError(context.keyManager, "NotAuthorised")
           .withArgs(
-            reentrancyContext.reentrantContract.address,
+            reentrancyContext.reentrantSigner.address,
             testCase.missingPermission
           );
       });
     });
 
-    it("should pass if the reentrant contract has the following permissions: REENTRANCY, ADDUNIVERSALRECEIVERDELEGATE", async () => {
+    it("should pass if the reentrant signer has the following permissions: REENTRANCY, ADDUNIVERSALRECEIVERDELEGATE", async () => {
       await loadTestCase(
         "ADDUNIVERSALRECEIVERDELEGATE",
         addUniversalReceiverDelegateTestCases.ValidCase,
         context,
-        reentrancyContext.reentrantContract.address,
-        reentrancyContext.reentrantContract.address
+        reentrancyContext.reentrantSigner.address,
+        reentrancyContext.batchReentarncyRelayer.address
       );
 
       await context.universalProfile
@@ -509,37 +481,27 @@ export const testERC725XExecuteToERC725XExecute = (
   });
 
   describe("when reentering and changing URD", () => {
-    let executeCalldata: {
-      operationType: number;
-      to: string;
-      value: number;
-      data: BytesLike;
-    };
-
     before(async () => {
-      const reentrantCall =
-        new LSP20ReentrantContract__factory().interface.encodeFunctionData(
-          "callThatReenters",
-          ["CHANGEUNIVERSALRECEIVERDELEGATE"]
-        );
-
-      executeCalldata = {
-        operationType: 0,
-        to: reentrancyContext.reentrantContract.address,
-        value: 0,
-        data: reentrantCall,
-      };
+      await generateBatchRelayPayload(
+        context.universalProfile,
+        context.keyManager,
+        "CHANGEUNIVERSALRECEIVERDELEGATE",
+        reentrancyContext.batchReentarncyRelayer,
+        reentrancyContext.reentrantSigner,
+        reentrancyContext.newControllerAddress,
+        reentrancyContext.newURDAddress
+      );
     });
 
     changeUniversalReceiverDelegateTestCases.NotAuthorised.forEach(
       (testCase) => {
-        it(`should revert if the reentrant contract has the following permission set: PRESENT - ${testCase.permissionsText}; MISSING - ${testCase.missingPermission};`, async () => {
+        it(`should revert if the reentrant signer has the following permission set: PRESENT - ${testCase.permissionsText}; MISSING - ${testCase.missingPermission};`, async () => {
           await loadTestCase(
             "CHANGEUNIVERSALRECEIVERDELEGATE",
             testCase,
             context,
-            reentrancyContext.reentrantContract.address,
-            reentrancyContext.reentrantContract.address
+            reentrancyContext.reentrantSigner.address,
+            reentrancyContext.batchReentarncyRelayer.address
           );
 
           await expect(
@@ -554,20 +516,20 @@ export const testERC725XExecuteToERC725XExecute = (
           )
             .to.be.revertedWithCustomError(context.keyManager, "NotAuthorised")
             .withArgs(
-              reentrancyContext.reentrantContract.address,
+              reentrancyContext.reentrantSigner.address,
               testCase.missingPermission
             );
         });
       }
     );
 
-    it("should pass if the reentrant contract has the following permissions: REENTRANCY, CHANGEUNIVERSALRECEIVERDELEGATE", async () => {
+    it("should pass if the reentrant signer has the following permissions: REENTRANCY, CHANGEUNIVERSALRECEIVERDELEGATE", async () => {
       await loadTestCase(
         "CHANGEUNIVERSALRECEIVERDELEGATE",
         changeUniversalReceiverDelegateTestCases.ValidCase,
         context,
-        reentrancyContext.reentrantContract.address,
-        reentrancyContext.reentrantContract.address
+        reentrancyContext.reentrantSigner.address,
+        reentrancyContext.batchReentarncyRelayer.address
       );
 
       await context.universalProfile
