+++ tests/Reentrancy/LSP20/reentrancyHelpers.ts
@@ -1,6 +1,6 @@
 import { ethers } from "hardhat";
 
-//types
+// types
 import { SignerWithAddress } from "@nomiclabs/hardhat-ethers/signers";
 import { BigNumber, BytesLike, Wallet } from "ethers";
 import {
@@ -13,7 +13,7 @@ import {
   SingleReentrancyRelayer__factory,
   BatchReentrancyRelayer__factory,
   UniversalProfile__factory,
-} from "../../../../types";
+} from "../../../types";
 
 // constants
 import {
@@ -21,10 +21,10 @@ import {
   ALL_PERMISSIONS,
   PERMISSIONS,
   CALLTYPE,
-} from "../../../../constants";
+} from "../../../constants";
 
 // setup
-import { LSP6TestContext } from "../../../utils/context";
+import { LSP6TestContext } from "../../utils/context";
 
 // helpers
 import {
@@ -33,8 +33,8 @@ import {
   LOCAL_PRIVATE_KEYS,
   signLSP6ExecuteRelayCall,
   encodeCompactBytesArray,
-} from "../../../utils/helpers";
-import { setupKeyManager } from "../../../utils/fixtures";
+} from "../../utils/helpers";
+import { setupKeyManager } from "../../utils/fixtures";
 
 // Complex permission as it has AllowedCalls
 export type TransferValueTestCase = {
